+++ packages/compile-swc/src/index.ts
@@ -1,7 +1,1 @@
-import swc from './swc-loader'
-export const compileType = 'swc'
-export const loader = swc
-export default {
-  compileType,
-  loader,
-}
+export {loader, compileType} from './loader'