+++ packages/compile-swc/src/loader.ts
@@ -1,0 +1,9 @@
+import type {ResovleConfig} from '@efox/emp'
+export const loader = (options: ResovleConfig) => {
+  // console.log('options', options)
+  return {
+    loader: require.resolve('./swc'),
+    options,
+  }
+}
+export const compileType = 'swc'