+++ packages/emp/src/cli/build.ts
@@ -31,8 +31,13 @@ class Build {
     }
 
     if (store.config.debug.clearLog) clearConsole()
-    logTag(`build mode ${store.config.mode}${this.isLib ? ' in [Library Mode] ' : ''}:`)
-    if (store.config.compile.compileType !== 'babel') logTag(`use ${store.config.compile.compileType}`, 'purple')
+    logTag(
+      `build mode ${store.config.mode}${this.isLib ? ' in [Library Mode] ' : ''} - ${
+        store.config.compile.compileType
+      } compile`,
+      store.config.compile.compileType !== 'babel' ? 'orange' : 'blue',
+    )
+    // if (store.config.compile.compileType !== 'babel') logTag(`use ${store.config.compile.compileType}`, 'purple')
     // await reporter.measureFileSizesBeforeBuild()
     webpack(this.config, (err: any, stats: any) => {
       if (err) {