+++ packages/emp/src/cli/dev.ts
@@ -14,8 +14,11 @@ class devServer {
   async setServer() {
     const config = getConfig()
     if (store.config.debug.clearLog) clearConsole()
-    logTag(`dev server running at:`)
-    if (store.config.compile.compileType !== 'babel') logTag(`use ${store.config.compile.compileType}`, 'purple')
+    logTag(
+      `dev server running - ${store.config.compile.compileType} compile`,
+      store.config.compile.compileType !== 'babel' ? 'orange' : 'blue',
+    )
+    // if (store.config.compile.compileType !== 'babel') logTag(`use ${store.config.compile.compileType}`, 'purple')
     //
     const compiler = webpack(config)
     this.server = new WebpackDevServer(config.devServer || {}, compiler)